refactor: Service인터페이스와 JDBCService를 일원화
