+++ src/main/java/me/kimihiqq/vouchermanagement/domain/customer/service/CustomerService.java
@@ -1,23 +1,51 @@
 package me.kimihiqq.vouchermanagement.domain.customer.service;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import me.kimihiqq.vouchermanagement.domain.customer.Customer;
 import me.kimihiqq.vouchermanagement.domain.customer.dto.CustomerDto;
+import me.kimihiqq.vouchermanagement.domain.customer.repository.CustomerRepository;
+import me.kimihiqq.vouchermanagement.domain.voucherwallet.service.VoucherWalletService;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
-public interface CustomerService {
+@Profile({"db", "test"})
+@Slf4j
+@RequiredArgsConstructor
+@Service
+public class CustomerService {
 
-    Customer createCustomer(CustomerDto customerDto);
-    List<Customer> listCustomers();
-    Optional<Customer> findCustomerById(UUID customerId);
-//    void deleteCustomerById(UUID customerId);
-//    void addVoucherToCustomer(UUID customerId, UUID voucherId);
-//    void removeVoucherFromCustomer(UUID customerId, UUID voucherId);
+    private final CustomerRepository customerRepository;
+    private final VoucherWalletService voucherWalletService;
 
-    void updateCustomerStatus(Customer customer);
 
-    List<Customer> findCustomersWithVoucher(UUID voucherId);
+    public Customer createCustomer(CustomerDto customerDto) {
+        Customer customer = customerDto.toCustomer();
+        return customerRepository.save(customer);
+    }
+
+
+    public List<Customer> listCustomers() {
+        return customerRepository.findAll();
+    }
+
+    public Optional<Customer> findCustomerById(UUID customerId) {
+        return customerRepository.findById(customerId);
+    }
+
+    public void updateCustomerStatus(Customer customer) {
+        customerRepository.update(customer);
+    }
+
+    public List<Customer> findCustomersWithVoucher(UUID voucherId) {
+        Set<UUID> customerIds = voucherWalletService.findCustomerIdsByVoucherId(voucherId);
+        List<Customer> customers = new ArrayList<>();
+        for(UUID id: customerIds){
+            findCustomerById(id).ifPresent(customers::add);
+        }
+        return customers;
+    }
 
 }
